requirements-dev.txt
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
coverage>=7.2.7
factory-boy>=3.3.0
faker>=18.13.0

# Development
black>=23.12.1
flake8>=7.0.0
isort>=5.13.2
mypy>=1.8.0
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=2.0.0 
waitress
pyinstaller